pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sota-data"
version = "0.1.0"
description = "SOTA Hackathon - Surgical Robotics Training Data"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "lerobot>=0.1.0",
    "rerun-sdk>=0.21.0",
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
    "torch>=2.0.0",
]

[project.optional-dependencies]
labeling = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
]

[project.scripts]
sota-visualize = "sota_data.visualize:main"

[tool.hatch.build.targets.wheel]
packages = ["src/sota_data"]

[tool.hatch.build.targets.sdist]
include = ["src/sota_data"]
